Allow float type for Circle radius parameter.
Previously only allowed int

+++ src/ansys/motorcad/core/geometry_shapes.py
@@ -258,7 +258,7 @@ class Circle(EntityList):
         Radius of the circle.
     """
 
-    def __init__(self, centre: Coordinate, radius: int):
+    def __init__(self, centre: Coordinate, radius: float):
         """Initialise Circle object."""
         super().__init__()
         self.append(
